@@ -13,23 +13,22 @@
 	},
 	"devDependencies": {
 		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"autoprefixer": "^10.4.20",
 		"mdsvex": "^0.12.3",
 		"phosphor-svelte": "^2.0.1",
 		"postcss": "^8.4.47",
 		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
-		"svelte": "^4.2.7",
+		"prettier-plugin-svelte": "^3.2.6",
+		"svelte": "^5.0.0",
 		"svelte-check": "^4.0.0",
 		"tailwindcss": "^3.4.14",
-		"typescript": "^5.0.0",
-		"vite": "^5.0.3"
+		"typescript": "^5.5.0",
+		"vite": "^5.4.4"
 	},
 	"type": "module",
 	"dependencies": {
-		"@splidejs/svelte-splide": "^0.2.9",
 		"bigger-picture": "^1.1.17",
 		"class-variance-authority": "^0.7.0",
 		"clsx": "^2.1.1",
