@@ -1,9 +1,13 @@
 <script>
-    import "../app.css";
+	import "../app.css";
 	import Footer from "./_components/Footer.svelte";
 	import Header from "./_components/Header.svelte";
+	/** @type {{children?: import('svelte').Snippet}} */
+	let { children } = $props();
 </script>
 
 
-<slot />
+
+
+{@render children?.()}
 <Footer />
