@@ -6,7 +6,11 @@
     import { fade } from "svelte/transition";
 
 
-    export let project: Project;
+    interface Props {
+        project: Project;
+    }
+
+    let { project }: Props = $props();
 
     console.log("This is Project Card")
 
@@ -19,7 +23,7 @@
     
 
 <div class="relative card" transition:fade={{ duration: 200 }}>
-    <button on:click={handleLike} class="flex w-fit p-1 absolute right-3 top-3 rounded-full bg-black/90 backdrop-blur-sm invert z-30">
+    <button onclick={handleLike} class="flex w-fit p-1 absolute right-3 top-3 rounded-full bg-black/90 backdrop-blur-sm invert z-30">
         <Heart  size={18} color="white" />
     </button>
 <a href={`/projects/${project.slug}`}>
