@@ -3,11 +3,11 @@
 </script>
 
 
-<section class="flex flex-col justify-end bg-gradient-to-b from-gray-50/20 via-gray-300/90 to-gray-300/80 h-64">  
+<section class="flex flex-col justify-end bg-gradient-to-b  from-gray-200/10 to-gray-100/90 h-64">  
     <div class="flex items-center justify-center gap-2 px-2 py-2">
             {#each SOCIALS as social}
                 <a href={social.href} target="_blank" class="bg-zinc-900 border border-zinc-900 transition-all duration-200 rounded-full p-1.5 gap-1 uppercase text-sm hover:invert hover:border-zinc-300">
-                    <svelte:component this={social.icon} size={24} weight="light" color="white" />
+                    <social.icon size={24} weight="light" color="white" />
                 </a>
             {/each}
     </div>
